# Speed-up common cases for deque.__init__()