Timings with patch:

```
$ ./python.exe -m timeit -r11 -s 'from collections import deque' 'deque()'
5000000 loops, best of 11: 77.2 nsec per loop
$ ./python.exe -m timeit -r11 -s 'from collections import deque' 'deque("abc")'
2000000 loops, best of 11: 157 nsec per loop
$ ./python.exe -m timeit -r11 -s 'from collections import deque' 'deque("abc", 2)'
2000000 loops, best of 11: 166 nsec per loop
```

Baseline timings:
```
$ ./python.exe -m timeit -r11 -s 'from collections import deque' 'deque()'
5000000 loops, best of 11: 81.4 nsec per loop
$ ./python.exe -m timeit -r11 -s 'from collections import deque' 'deque("abc")'
2000000 loops, best of 11: 171 nsec per loop
$ ./python.exe -m timeit -r11 -s 'from collections import deque' 'deque("abc", 2)'
2000000 loops, best of 11: 181 nsec per loop
```